refactor: smart and dumb components
