+++ firebase-data/auth_export/accounts.json
@@ -1,1 +1,1 @@
-{"kind":"identitytoolkit#DownloadAccountResponse","users":[{"localId":"E2Z4H6mXwTMcPc3sHpCjmBSfRmK3","lastLoginAt":"1698244886477","emailVerified":false,"email":"email@email.com","salt":"fakeSaltiFh2hMb2StXH6Ul2Mbk4","passwordHash":"fakeHash:salt=fakeSaltiFh2hMb2StXH6Ul2Mbk4:password=password","passwordUpdatedAt":1698244886477,"validSince":"1698244886","createdAt":"1698244886477","providerUserInfo":[{"providerId":"password","email":"email@email.com","federatedId":"email@email.com","rawId":"email@email.com"}],"lastRefreshAt":"2023-10-25T14:41:26.479Z"},{"localId":"lk2OK6qG44hzx0ywSRP8QC20OePw","lastLoginAt":"1698244929051","emailVerified":false,"email":"email2@email.com","salt":"fakeSaltpN2XJuSvdg93G44Gz9bN","passwordHash":"fakeHash:salt=fakeSaltpN2XJuSvdg93G44Gz9bN:password=password2","passwordUpdatedAt":1698244929051,"validSince":"1698244929","createdAt":"1698244929051","providerUserInfo":[{"providerId":"password","email":"email2@email.com","federatedId":"email2@email.com","rawId":"email2@email.com"}],"lastRefreshAt":"2023-10-25T14:42:09.051Z"}]}
+{"kind":"identitytoolkit#DownloadAccountResponse","users":[{"localId":"E2Z4H6mXwTMcPc3sHpCjmBSfRmK3","createdAt":"1698244886477","lastLoginAt":"1698244886477","passwordHash":"fakeHash:salt=fakeSaltiFh2hMb2StXH6Ul2Mbk4:password=password","salt":"fakeSaltiFh2hMb2StXH6Ul2Mbk4","passwordUpdatedAt":1698324640414,"providerUserInfo":[{"providerId":"password","email":"email@email.com","federatedId":"email@email.com","rawId":"email@email.com"}],"validSince":"1698324640","email":"email@email.com","emailVerified":false,"disabled":false},{"localId":"lk2OK6qG44hzx0ywSRP8QC20OePw","createdAt":"1698244929051","lastLoginAt":"1698244929051","passwordHash":"fakeHash:salt=fakeSaltpN2XJuSvdg93G44Gz9bN:password=password2","salt":"fakeSaltpN2XJuSvdg93G44Gz9bN","passwordUpdatedAt":1698324640414,"providerUserInfo":[{"providerId":"password","email":"email2@email.com","federatedId":"email2@email.com","rawId":"email2@email.com"}],"validSince":"1698324640","email":"email2@email.com","emailVerified":false,"disabled":false}]}
