+++ package.json
@@ -17,7 +17,8 @@
     "type": "tsc --project tsconfig.json",
     "lint": "eslint . --ext js,jsx,ts,tsx",
     "lint:fix": "eslint . --ext js,jsx,ts,tsx --fix",
-    "emu:start": "firebase emulators:start --import=./firebase-data --export-on-exit=./firebase-data",
+    "emu:export": "firebase emulators:export",
+    "emu:start": "firebase emulators:start --import=./firebase-data --export-on-exit",
     "on:dev": "npm-run-all --parallel typesafe-i18n dev",
     "onfire:dev": "npm-run-all --parallel typesafe-i18n dev emu:start",
     "dev": "rimraf dev-dist && DEBUG=vite-plugin-pwa SW_DEV=true vite --force",
