+++ src/modules/shared/components/dumb/SvgIcon/SvgIcon.atom.tsx
