+++ src/modules/shared/components/smart/RouteErrorBoundary/RouteErrorBoundary.template.tsx
