+++ src/modules/todo/routes/todos.route.tsx
@@ -1,5 +1,5 @@
-import NavbarWrapper from '@shared/components/templates/NavbarWrapper/NavbarWrapper.template';
-import RouteErrorBoundary from '@shared/components/templates/RouteErrorBoundary/RouteErrorBoundary.template';
+import NavbarWrapper from '@shared/components/dumb/NavbarWrapper/NavbarWrapper.template';
+import RouteErrorBoundary from '@shared/components/smart/RouteErrorBoundary/RouteErrorBoundary.template';
 import { RouteObject } from 'react-router-dom';
 
 export const todosId = {
